+++ src/components/Button/Button.js
@@ -1,21 +1,42 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-const Button = ({ type, category, label, onClick }) => (
-  /* eslint-disable-next-line react/button-has-type */
-  <button type={type} className={'uk-button uk-button-' + category} onClick={onClick}>{label}</button>
-);
+const Button = ({ category, label, colors, onClick }) => {
+  const getStyle = () => {
+    if (colors) {
+      return {
+        backgroundColor: colors.background,
+        color: colors.text
+      };
+    }
+    return {};
+  };
+
+  return (
+    <button
+      type="button"
+      className={'uk-button uk-button-' + category}
+      style={getStyle()}
+      onClick={onClick}
+    >
+      {label}
+    </button>
+  );
+};
 
 Button.propTypes = {
-  type: PropTypes.string,
   category: PropTypes.string,
   label: PropTypes.string.isRequired,
+  colors: PropTypes.shape({
+    background: PropTypes.string,
+    text: PropTypes.string,
+  }),
   onClick: PropTypes.func.isRequired
 };
 
 Button.defaultProps = {
-  type: 'button',
-  category: 'default'
+  category: 'default',
+  colors: null
 };
 
 export default Button;